half day leave and delete 	modified:   src/main/java/com/concretepage/controller/CustomerProgramController.java 	modified:   src/main/java/com/concretepage/controller/EmployeeLeaveController.java 	modified:   src/main/java/com/concretepage/dao/ICustomerProgramDao.java 	modified:   src/main/java/com/concretepage/dao/IEmployeeLeaveBalanceDao.java 	modified:   src/main/java/com/concretepage/dao/impl/CustomerDAOImpl.java 	modified:   src/main/java/com/concretepage/dao/impl/CustomerProgramDaoImpl.java 	modified:   src/main/java/com/concretepage/dao/impl/DepartmentDAOImpl.java 	modified:   src/main/java/com/concretepage/dao/impl/EmployeeLeaveBalanceDao.java 	modified:   src/main/java/com/concretepage/dao/impl/ProjectDAOImpl.java 	modified:   src/main/java/com/concretepage/dao/impl/RoleDAOImpl.java 	modified:   src/main/java/com/concretepage/dao/impl/TaskDAOImpl.java 	modified:   src/main/java/com/concretepage/entity/Customer.java 	modified:   src/main/java/com/concretepage/entity/CustomerProgram.java 	modified:   src/main/java/com/concretepage/entity/Department.java 	modified:   src/main/java/com/concretepage/entity/EmployeeLeave.java 	modified:   src/main/java/com/concretepage/entity/EmployeeLeaveBalance.java 	modified:   src/main/java/com/concretepage/entity/LeaveMaster.java 	modified:   src/main/java/com/concretepage/entity/Project.java 	modified:   src/main/java/com/concretepage/entity/Role.java 	modified:   src/main/java/com/concretepage/entity/Task.java 	modified:   src/main/java/com/concretepage/service/EmployeeService.java 	modified:   src/main/java/com/concretepage/service/ICustomerProgramService.java 	modified:   src/main/java/com/concretepage/service/ILeaveService.java 	modified:   src/main/java/com/concretepage/service/impl/CustomerProgramServiceImpl.java 	modified:   src/main/java/com/concretepage/service/impl/LeaveService.java 	modified:   src/main/java/com/concretepage/utils/HRUtils.java 	modified:   src/main/resources/SQlChanges.txt

+++ spring-boot-demo/src/main/java/com/concretepage/controller/EmployeeLeaveController.java
@@ -96,7 +96,7 @@ public class EmployeeLeaveController {
 		leave.setLeaveType(leaveType);
 		leave.setFromDate(fromDate);
 		leave.setToDate(toDate);
-		leave.setNoOfDays(Integer.parseInt(noOfDays));
+		leave.setNoOfDays(Float.parseFloat(noOfDays));
 		leave.setFromHours(0);
 		leave.setToHours(0);
 		leave.setComments(comments);
@@ -124,7 +124,7 @@ public class EmployeeLeaveController {
 			@PathVariable("year") String year, 
 			@PathVariable("noOfDays") String noOfDays
 			) throws HRException {
-		leaveService.incrementLeaveBalanceForAllEmployees(Integer.parseInt(year),leaveType,Integer.parseInt(noOfDays));
+		leaveService.incrementLeaveBalanceForAllEmployees(Integer.parseInt(year),leaveType,Float.parseFloat(noOfDays));
 		
 		return new ResponseEntity<Void>( HttpStatus.OK);
 	}
