+++ spring-boot-demo/src/main/java/com/concretepage/dao/IEmployeeLeaveBalanceDao.java
@@ -6,17 +6,17 @@ import com.concretepage.entity.EmployeeLeaveBalance;
 
 public interface IEmployeeLeaveBalanceDao {
 	
-	int incrementPendingApprovalLeaveBalance(Integer employeeId, Integer year, String leaveType, Integer noOfDay);
+	int incrementPendingApprovalLeaveBalance(Integer employeeId, Integer year, String leaveType, Float noOfDay);
 	
-	int decreaseEligibleLeaveBalance(Integer employeeId, Integer year, String leaveType, Integer noOfDay);
+	int decreaseEligibleLeaveBalance(Integer employeeId, Integer year, String leaveType, Float noOfDay);
 	
 	List<EmployeeLeaveBalance> getLeaveBalance(Integer employeeId, Integer year);
 	
-	int insertLeaveBalance(Integer employeeId, Integer year, String leaveType, Integer noOfDayPending, Integer noOfDayEligible);
+	int insertLeaveBalance(Integer employeeId, Integer year, String leaveType, Float noOfDayPending, Float noOfDayEligible);
 	
-	int incrementLeaveBalance(Integer employeeId, Integer year, String leaveType, Integer noOfDayPendingApproval, Integer noofDayEligible);
+	int incrementLeaveBalance(Integer employeeId, Integer year, String leaveType, Float noOfDayPendingApproval, Float noofDayEligible);
 	
-	public int incrementLeaveBalanceForAllEmployees( Integer year, String leaveType, Integer noofDayEligible);
+	public int incrementLeaveBalanceForAllEmployees( Integer year, String leaveType, Float noofDayEligible);
 	
 	
 }
